refactor: rename configurations and listener
