+++ backend/EntryService/src/main/java/EntryService/Configuration/EntryServiceConfiguration.java
@@ -6,7 +6,7 @@ import org.springframework.context.annotation.Configuration;
 import org.springframework.web.client.RestTemplate;
 
 @Configuration
-public class MyConfiguration {
+public class EntryServiceConfiguration {
     @Bean
     @LoadBalanced
     public RestTemplate restTemplate() {
