+++ backend/ManagementService/src/main/java/ManagementService/Configuration/ManagementServiceConfiguration.java
@@ -1,4 +1,4 @@
-package PowerPlantPackage.Configuration;
+package ManagementService.Configuration;
 
 import org.springframework.cloud.client.loadbalancer.LoadBalanced;
 import org.springframework.context.annotation.Bean;
@@ -6,10 +6,10 @@ import org.springframework.context.annotation.Configuration;
 import org.springframework.web.client.RestTemplate;
 
 @Configuration
-public class MyConfiguration {
-    @LoadBalanced
+public class ManagementServiceConfiguration {
     @Bean
-    RestTemplate restTemplate(){
+    @LoadBalanced
+    public RestTemplate restTemplate() {
         return new RestTemplate();
     }
 }
