+++ backend/PowerPlantService/src/main/java/PowerPlantPackage/Configuration/PowerPlantServiceConfiguration.java
@@ -1,4 +1,4 @@
-package ManagementService.Configuration;
+package PowerPlantPackage.Configuration;
 
 import org.springframework.cloud.client.loadbalancer.LoadBalanced;
 import org.springframework.context.annotation.Bean;
@@ -6,10 +6,10 @@ import org.springframework.context.annotation.Configuration;
 import org.springframework.web.client.RestTemplate;
 
 @Configuration
-public class MyConfiguration {
-    @Bean
+public class PowerPlantServiceConfiguration {
     @LoadBalanced
-    public RestTemplate restTemplate() {
+    @Bean
+    RestTemplate restTemplate(){
         return new RestTemplate();
     }
 }
