+++ backend/StatisticsService/src/main/java/StatisticsService/Configuration/StatisticsServiceConfiguration.java
@@ -6,7 +6,7 @@ import org.springframework.context.annotation.Configuration;
 import org.springframework.web.client.RestTemplate;
 
 @Configuration
-public class MyConfiguration {
+public class StatisticsServiceConfiguration {
     @Bean
     @LoadBalanced
     public RestTemplate restTemplate() {
