+++ backend/UserService/src/main/java/UserService/Configuration/UserServiceConfiguration.java
@@ -6,7 +6,7 @@ import org.springframework.context.annotation.Configuration;
 import org.springframework.web.client.RestTemplate;
 
 @Configuration
-public class MyConfiguration {
+public class UserServiceConfiguration {
     @Bean
     @LoadBalanced
     public RestTemplate restTemplate() {
